// apps/web/src/app/features/inventory/modules/hospitals/types/hospitals.types.ts
export interface Hospital {
  id: number;
  hospital_code: string;
  hospital_name: string;
  hospital_type?: string | null;
  province?: string | null;
  region?: string | null;
  is_active?: boolean | null;
  created_at?: string | null;
  updated_at?: string | null;
}